requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
google-api-core==2.4.0
google-auth==2.6.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.2
google-cloud-logging==2.7.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.10.1
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.19
urlfetch==1.2.2
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.8.0
